test: compare AM62L BGA fanout solvers
