+++ tests/repros/repro-am62l-lpddr4-two-bus-bga-fanout.test.tsx
@@ -1,0 +1,11 @@
+import { test } from "bun:test"
+import { createBgaFanoutAlgorithm } from "tests/fixtures/create-bga-fanout-algorithm"
+import { renderAm62lLpddr4TwoBusFanout } from "tests/fixtures/create-am62l-lpddr4-two-bus-fanout"
+
+test("routes two DDR byte buses with bga-fanout-solver", async () => {
+  await renderAm62lLpddr4TwoBusFanout({
+    fanoutAlgorithmFn: createBgaFanoutAlgorithm,
+    fanoutSolverLabel: "@tscircuit/bga-fanout-solver",
+    snapshotPath: import.meta.path,
+  })
+}, 300_000)
